+++ frontend/src/pages/event-groups/new.tsx
@@ -1,34 +1,23 @@
 import {
   Button,
   Input,
-  Text,
   Heading,
   Box,
-  Flex,
-  NumberInput,
-  NumberInputField,
-  NumberInputStepper,
-  NumberIncrementStepper,
-  NumberDecrementStepper,
   Spinner,
   Alert,
   AlertIcon,
   AlertTitle,
   AlertDescription,
   Container,
-  IconButton,
-  Icon,
 } from "@chakra-ui/react";
 import type { NextPage } from "next";
-import { useState, useCallback, useEffect } from "react";
 import { Web3Storage } from "web3.storage";
 import { useCreateEventGroup, INFTImage } from "../../hooks/useEventManager";
-import ImageSelectorWithPreview from "../../components/ImageSelectorWithPreview";
 import LoginRequired from "../../components/atoms/web3/LoginRequired";
 import { useLocale } from "../../hooks/useLocale";
 import { Controller, useFieldArray, useForm } from "react-hook-form";
-import { CloseIcon } from "@chakra-ui/icons";
 import ErrorMessage from "../../components/atoms/form/ErrorMessage";
+import NFTAttributesForm from "../../components/organisms/NFTAttributesForm";
 
 if (!process.env.NEXT_PUBLIC_WEB3_STORAGE_KEY) {
   throw new Error("WEB3_STORAGE_KEY is required");
@@ -44,7 +33,7 @@ const renameFile = (file: File, newFilename: string) => {
   return new File([file], newFilename, { type, lastModified });
 };
 
-interface FormData {
+export interface EventGroupFormData {
   groupName: string;
   nfts: {
     description: string;
@@ -60,8 +49,9 @@ const NewEventGroupPage: NextPage = () => {
   const {
     control,
     handleSubmit,
-    formState: { isSubmitting, isSubmitSuccessful },
-  } = useForm<FormData>({
+    watch,
+    formState: { isSubmitting },
+  } = useForm<EventGroupFormData>({
     defaultValues: {
       groupName: "",
       nfts: [
@@ -73,20 +63,20 @@ const NewEventGroupPage: NextPage = () => {
         {
           description: "",
           fileObject: null,
-          requiredParticipateCount: 5,
+          requiredParticipateCount: 3,
         },
         {
           description: "",
           fileObject: null,
-          requiredParticipateCount: 10,
+          requiredParticipateCount: 5,
         },
       ],
     },
   });
 
-  const { fields: nfts, remove } = useFieldArray({ control, name: "nfts" });
+  const { remove, append } = useFieldArray({ control, name: "nfts" });
 
-  const uploadImagesToIpfs = async (nfts: FormData["nfts"]) => {
+  const uploadImagesToIpfs = async (nfts: EventGroupFormData["nfts"]) => {
     console.log("Starting to upload images to IPFS...");
     const renamedFiles = nfts.map(
       ({ fileObject, description, requiredParticipateCount }) => ({
@@ -113,7 +103,7 @@ const NewEventGroupPage: NextPage = () => {
     return { rootCid, renamedFiles };
   };
 
-  const submit = async (data: FormData) => {
+  const submit = async (data: EventGroupFormData) => {
     const uploadResult = await uploadImagesToIpfs(data.nfts);
     if (!uploadResult) {
       console.error("uploading error");
@@ -169,99 +159,14 @@ const NewEventGroupPage: NextPage = () => {
               <Heading as="h2" fontSize="xl" my={4}>
                 NFTs
               </Heading>
-              <Box>
-                {nfts.map((nft, index) => (
-                  <Flex
-                    key={index}
-                    w="full"
-                    flexDirection={{ base: "column", md: "row" }}
-                    mb={10}
-                    position="relative"
-                  >
-                    <Box flexBasis="300px" flexShrink="0" minH="300px" mb={3}>
-                      <Controller
-                        control={control}
-                        name={`nfts.${index}.fileObject`}
-                        rules={{ required: "NFT Image is required" }}
-                        render={({ field: { onChange } }) => (
-                          <>
-                            <ImageSelectorWithPreview
-                              onChangeData={(newFile) => {
-                                onChange(newFile);
-                              }}
-                            />
-                          </>
-                        )}
-                      />
-                    </Box>
-                    <Box ml={{ md: 5, base: 0 }}>
-                      <Box>
-                        <Text>{t.NFT_DESC}</Text>
-                        <Controller
-                          control={control}
-                          name={`nfts.${index}.description`}
-                          rules={{
-                            required: "NFT description is required",
-                          }}
-                          render={({ field: { onChange, value } }) => (
-                            <Input
-                              variant="outline"
-                              mb={4}
-                              value={value}
-                              onChange={(e) => {
-                                onChange(e.target.value);
-                              }}
-                            />
-                          )}
-                        />
-                      </Box>
-                      <Box mt={2}>
-                        {index > 0 ? (
-                          <>
-                            <Text>{t.TIMES_PARTICIPATE}</Text>
-                            <Controller
-                              control={control}
-                              name={`nfts.${index}.requiredParticipateCount`}
-                              rules={{
-                                required: "Times participation is required",
-                                min: 1,
-                              }}
-                              render={({ field: { onChange, value } }) => (
-                                <NumberInput
-                                  value={value}
-                                  min={0}
-                                  onChange={(__, num) => {
-                                    onChange(num);
-                                  }}
-                                >
-                                  <NumberInputField />
-                                  <NumberInputStepper>
-                                    <NumberIncrementStepper />
-                                    <NumberDecrementStepper />
-                                  </NumberInputStepper>
-                                </NumberInput>
-                              )}
-                            />
-                          </>
-                        ) : (
-                          <Text>{t.NFT_DEFAULT}</Text>
-                        )}
-                      </Box>
-                    </Box>
-                    {index !== 0 && (
-                      <IconButton
-                        position="absolute"
-                        right="0"
-                        top="0"
-                        borderRadius="full"
-                        aria-label=""
-                        icon={<Icon as={CloseIcon} color="mint.primary" />}
-                        onClick={() => remove(index)}
-                      />
-                    )}
-                  </Flex>
-                ))}
-              </Box>
+              {control && (
+                <NFTAttributesForm
+                  control={control}
+                  nfts={watch("nfts")}
+                  append={append}
+                  remove={remove}
+                />
+              )}
               <Box mt={8} mb={10}>
                 <Button
                   disabled={isSubmitting || loading}
